reorg k8s module & add function coords env var
ASML_FUNCTION_COORDINATES must be set in single-function environments for the Hyper runtime
